requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
kafka-python==2.3.0
